ResultFormat and PreEncoded
`ResultFormat.raw` has been removed, as its completely superfluous for when the `ResponseType == Raw` (this is a breaking change but all a dev must do is remove the `var resultFormat: ResultFormat { return .raw }` or however they've implemented it.

Also adds a new `PreEncoded` protocol. When a Restable conforms to this protocol, Restivus will use whatever the value of the `data` field is for the HTTPRequest body.